make `is_platform_object_same_origin` safe and move it out from `DomHelpers` trait
